FPLA-2513: Add email template testing support - move packages
